:recycle: refactor: 패키지 이동 - jdbc 공통 사용으로 core 모듈로 이동
