--- semo-batch/src/main/resources/sql/batch-queries.properties
@@ -1,53 +1,0 @@
-updateDeviceStatus=\
-  UPDATE DEVICES SET STATUS = ? \
-  WHERE DEVICE_ID = ?
-selectSessionData=\
-  SELECT s.SID, "SERIAL#", s.USERNAME, COMMAND, COMMAND_NAME, s.STATUS, SCHEMANAME, OSUSER, PROCESS, MACHINE, PORT, TERMINAL, \
-         s.PROGRAM, TYPE, s.SQL_ID, s.SQL_EXEC_START, s.SQL_EXEC_ID, SQL_TEXT, s.MODULE, s.ACTION, LOGON_TIME, LAST_CALL_ET, \
-         FAILED_OVER, BLOCKING_SESSION_STATUS, EVENT, WAIT_CLASS, STATE, WAIT_TIME_MICRO, TIME_REMAINING_MICRO, s.SERVICE_NAME \
-  FROM v$session s \
-  LEFT OUTER JOIN v$sql_monitor m ON s.sql_id = m.sql_id AND s.sql_exec_start = m.sql_exec_start AND s.sql_exec_id = m.sql_exec_id \
-  LEFT OUTER JOIN v$sqlcommand c ON s.command = c.command_type
-insertSessionData=\
-  INSERT INTO SESSION_DATA ( COLLECTED_AT, SID, DEVICE_ID, \
-    SERIAL, USERNAME, COMMAND, COMMAND_NAME, STATUS, SCHEMANAME, OSUSER, PROCESS, MACHINE, PORT, TERMINAL, \
-    PROGRAM, TYPE, SQL_ID, SQL_EXEC_START, SQL_EXEC_ID, SQL_TEXT, MODULE, ACTION, LOGON_TIME, LAST_CALL_ET, \
-    FAILED_OVER, BLOCKING_SESSION_STATUS, EVENT, WAIT_CLASS, STATE, WAIT_TIME_MICRO, TIME_REMAINING_MICRO, SERVICE_NAME) \
-  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-selectMetricData=\
-  SELECT \
-    (SELECT COUNT(*) FROM v$session) AS TOTAL_SESSION_COUNT, \
-    (SELECT COUNT(*) FROM v$session WHERE STATUS = 'ACTIVE') AS ACTIVE_SESSION_COUNT, \
-    (SELECT LISTAGG(USERNAME || ':' || SESSION_CNT, ', ') \
-     FROM \
-        (SELECT USERNAME, COUNT(*) AS SESSION_CNT \
-         FROM v$session \
-         WHERE USERNAME IS NOT NULL \
-         GROUP BY USERNAME\
-    )) AS SESSION_COUNT_GROUP_BY_USER, \
-    (SELECT LISTAGG(COMMAND_NAME || ':' || COMMAND_CNT, ', ') \
-     FROM \
-        (SELECT c.COMMAND_NAME, COUNT(*) AS COMMAND_CNT \
-         FROM (SELECT COMMAND FROM v$session WHERE COMMAND != 0) s JOIN v$sqlcommand c ON s.COMMAND = c.COMMAND_TYPE \
-         WHERE COMMAND_NAME IS NOT NULL GROUP BY c.COMMAND_NAME\
-        )) AS SESSION_COUNT_GROUP_BY_COMMAND, \
-    (SELECT LISTAGG(MACHINE || ':' || MACHINE_CNT, ', ') \
-     FROM \
-        (SELECT MACHINE, COUNT(*) AS MACHINE_CNT \
-         FROM v$session \
-         GROUP BY MACHINE\
-    )) AS SESSION_COUNT_GROUP_BY_MACHINE, \
-    (SELECT LISTAGG(TYPE || ':' || TYPE_CNT, ', ') \
-     FROM \
-        (SELECT TYPE, COUNT(*) AS TYPE_CNT \
-         FROM v$session \
-         GROUP BY TYPE\
-    )) AS SESSION_COUNT_GROUP_BY_TYPE,\
-    (SELECT COUNT(*) FROM v$session WHERE BLOCKING_SESSION_STATUS = 'VALID') AS BLOCKING_SESSION_COUNT, \
-    (SELECT COUNT(*) FROM v$session WHERE STATE = 'WAITING' AND WAIT_CLASS != 'Idle') AS WAIT_SESSION_COUNT \
-  FROM DUAL
-insertMonitoringMetric=\
-  INSERT INTO MONITORING_METRICS (COLLECTED_AT, DEVICE_ID, \
-    TOTAL_SESSION_COUNT, ACTIVE_SESSION_COUNT, SESSION_COUNT_GROUP_BY_USER, SESSION_COUNT_GROUP_BY_COMMAND, \
-    SESSION_COUNT_GROUP_BY_MACHINE, SESSION_COUNT_GROUP_BY_TYPE, BLOCKING_SESSION_COUNT, WAIT_SESSION_COUNT) \
-  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
